security improvement - need admin access to delete all projects and settings
git-svn-id: https://svn.fao.org/projects/ruralinvest/RIV4/trunk@66821 d7fa552c-c32d-11e1-9f2d-6d39f2a56844

+++ RuralInvest4/src/main/java/riv/web/controller/MainController.java
@@ -93,8 +93,11 @@ public class MainController {
 	 * Used by testing framework
 	 */
 	@RequestMapping("help/deleteAll")
-	public String deleteAll(@RequestParam String type, @RequestParam String ig) {
-		dataService.deleteAll(type.equals("project"), ig.equals("true"));
+	public String deleteAll(@RequestParam String type, @RequestParam String ig, HttpServletRequest request) {
+		User u = (User)request.getAttribute("user");
+		if (u.isAdministrator()) {
+			dataService.deleteAll(type.equals("project"), ig.equals("true"));
+		}
 		return "redirect:../home";
 	}
 	
@@ -102,9 +105,12 @@ public class MainController {
 	 * Used by testing framework
 	 */
 	@RequestMapping("help/deleteAllAppConfigs")
-	public String deleteAllAppConfigs() {
-		dataService.deleteAllAppConfigs();
-		rivConfig.reload();
+	public String deleteAllAppConfigs(HttpServletRequest request) {
+		User u = (User)request.getAttribute("user");
+		if (u.isAdministrator()) {
+			dataService.deleteAllAppConfigs();
+			rivConfig.reload();
+		}
 		return "redirect:../home";
 	}
